# Conflicts:
#	acm-standard-applications/acm-law-enforcement/src/main/webapp/resources/services/object/object-lookup.client.service.js
#	acm-tool-integrations/acm-pdf-utilities/src/main/java/com/armedia/acm/pdf/service/PdfServiceImpl.java


Former-commit-id: 635b021ba5bc817263878f42a0ae92aa7bd1d6f6

+++ acm-services/acm-service-wopi/src/main/java/com/armedia/acm/services/wopi/api/WopiFilesApiController.java
@@ -29,10 +29,10 @@ package com.armedia.acm.services.wopi.api;
 
 import com.armedia.acm.core.exceptions.AcmObjectNotFoundException;
 import com.armedia.acm.core.exceptions.AcmUserActionFailedException;
+import com.armedia.acm.services.dataaccess.service.impl.ArkPermissionEvaluator;
 import com.armedia.acm.services.wopi.model.WopiFileInfo;
 import com.armedia.acm.services.wopi.model.WopiLockInfo;
 import com.armedia.acm.services.wopi.service.WopiAcmService;
-
 import org.mule.api.MuleException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -40,14 +40,13 @@ import org.springframework.core.io.InputStreamResource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
+import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import java.io.IOException;
@@ -57,25 +56,35 @@ import java.io.IOException;
 public class WopiFilesApiController
 {
     private WopiAcmService wopiService;
+    private ArkPermissionEvaluator permissionEvaluator;
     private static final Logger log = LoggerFactory.getLogger(WopiFilesApiController.class);
 
-    @PreAuthorize("hasPermission(#id, 'FILE', 'read|group-read|write|group-write')")
     @RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
     public ResponseEntity<WopiFileInfo> getFileInfo(@PathVariable Long id, Authentication authentication)
     {
         log.info("Getting file info for id [{}] per user [{}]", id, authentication.getName());
-        return wopiService.getFileInfo(id, authentication)
+        boolean hasPermission = isActionAllowed(authentication, "read|group-read|write|group-write", id);
+        if (!hasPermission)
+        {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+        return wopiService.getFileInfo(id)
                 .map(it -> new ResponseEntity<>(it, HttpStatus.OK))
                 .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
-    @PreAuthorize("hasPermission(#id, 'FILE', 'read|group-read|write|group-write')")
     @RequestMapping(value = "/{id}/contents", method = RequestMethod.GET, produces = MediaType.APPLICATION_OCTET_STREAM_VALUE)
     @ResponseBody
     public ResponseEntity<InputStreamResource> getFileContents(@PathVariable Long id, Authentication authentication)
     {
         log.info("Getting contents for file with id [{}] per user [{}]", id, authentication.getName());
+
+        boolean hasPermission = isActionAllowed(authentication, "read|group-read|write|group-write", id);
+        if (!hasPermission)
+        {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         try
         {
             InputStreamResource fileContents = wopiService.getFileContents(id);
@@ -88,7 +97,6 @@ public class WopiFilesApiController
         }
     }
 
-    @PreAuthorize("hasPermission(#id, 'FILE', 'write|group-write')")
     @RequestMapping(value = "/{id}/contents", method = RequestMethod.POST)
     @ResponseBody
     public ResponseEntity putFile(@PathVariable Long id,
@@ -96,6 +104,12 @@ public class WopiFilesApiController
             Authentication authentication)
     {
         log.info("Put file [{}] per user [{}]", id, authentication.getName());
+        boolean hasPermission = isActionAllowed(authentication, "write|group-write", id);
+        if (!hasPermission)
+        {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         try
         {
             wopiService.putFile(id, resource, authentication);
@@ -113,43 +127,60 @@ public class WopiFilesApiController
         }
     }
 
-    @PreAuthorize("hasPermission(#id, 'FILE', 'write|group-write')")
     @RequestMapping(value = "/{id}/lock", method = RequestMethod.PUT)
     @ResponseBody
     public ResponseEntity<WopiLockInfo> lockFile(@PathVariable Long id, Authentication authentication)
     {
         log.info("Lock file [{}] per user [{}]", id, authentication.getName());
+        boolean hasPermission = isActionAllowed(authentication, "write|group-write", id);
+
+        if (!hasPermission)
+        {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return new ResponseEntity<>(wopiService.lock(id, authentication), HttpStatus.OK);
     }
 
-    @PreAuthorize("hasPermission(#id, 'FILE', 'write|group-write')")
     @RequestMapping(value = "/{id}/lock", method = RequestMethod.GET)
     @ResponseBody
     public ResponseEntity<WopiLockInfo> getLock(@PathVariable Long id, Authentication authentication)
     {
         log.info("Get lock for file [{}] per user [{}]", id, authentication.getName());
+        boolean hasPermission = isActionAllowed(authentication, "write|group-write", id);
+        if (!hasPermission)
+        {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return new ResponseEntity<>(wopiService.getSharedLock(id), HttpStatus.OK);
     }
 
-    @PreAuthorize("hasPermission(#id, 'FILE', 'write|group-write')")
     @RequestMapping(value = "/{id}/lock/{lockId}", method = RequestMethod.DELETE)
     @ResponseBody
     public ResponseEntity<WopiLockInfo> unlock(@PathVariable Long id, @PathVariable Long lockId, Authentication authentication)
     {
         log.info("Unlock file [{}] per user [{}]", id, authentication.getName());
+        boolean hasPermission = isActionAllowed(authentication, "write|group-write", id);
+        if (!hasPermission)
+        {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         return new ResponseEntity<>(wopiService.unlock(id, lockId, authentication), HttpStatus.OK);
     }
 
-    @PreAuthorize("hasPermission(#id, 'FILE', 'write|group-write')")
     @RequestMapping(value = "/{id}/rename", method = RequestMethod.PUT, produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseBody
-    public ResponseEntity renameFile(@PathVariable Long id, @RequestParam String name,
+    public ResponseEntity renameFile(@PathVariable Long id, @RequestBody MultiValueMap<String, String> body,
             Authentication authentication)
     {
         log.info("Rename file [{}] per user [{}]", id, authentication.getName());
+        boolean hasPermission = isActionAllowed(authentication, "write|group-write", id);
+        if (!hasPermission)
+        {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         try
         {
-            wopiService.renameFile(id, name);
+            wopiService.renameFile(id, body.getFirst("name"));
             return new ResponseEntity<>(HttpStatus.OK);
         }
         catch (AcmObjectNotFoundException e)
@@ -164,12 +195,17 @@ public class WopiFilesApiController
         }
     }
 
-    @PreAuthorize("hasPermission(#id, 'FILE', 'write|group-write')")
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     @ResponseBody
     public ResponseEntity deleteFile(@PathVariable Long id, Authentication authentication)
     {
         log.info("Delete file [{}] per user [{}]", id, authentication.getName());
+        boolean hasPermission = isActionAllowed(authentication, "write|group-write", id);
+        if (!hasPermission)
+        {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         try
         {
             wopiService.deleteFile(id);
@@ -187,8 +223,18 @@ public class WopiFilesApiController
         }
     }
 
+    private boolean isActionAllowed(Authentication authentication, String action, Long targetId)
+    {
+        return permissionEvaluator.hasPermission(authentication, targetId, "FILE", action);
+    }
+
     public void setWopiService(WopiAcmService wopiService)
     {
         this.wopiService = wopiService;
+    }
+
+    public void setPermissionEvaluator(ArkPermissionEvaluator permissionEvaluator)
+    {
+        this.permissionEvaluator = permissionEvaluator;
     }
 }
